requirements.txt
python-telegram-bot == 12.3.0